monthly_report_bot_link_pack/requirements_v1_1.txt
# 月报机器人 v1.1 依赖包
# 基于需求说明书实现

# 核心依赖
requests>=2.31.0
PyYAML>=6.0.1
pytz>=2023.3
flask>=2.3.3

# WebSocket支持
websockets>=12.0
aiohttp>=3.9.0

# 异步支持
asyncio-mqtt>=0.16.1

# 数据处理
pandas>=2.1.0
numpy>=1.24.0

# 图表生成
matplotlib>=3.7.0
seaborn>=0.12.0

# 日志和监控
structlog>=23.2.0
prometheus-client>=0.19.0

# 智能交互
jieba>=0.42.1
langdetect>=1.0.9

# 安全和加密
cryptography>=41.0.0
hmac>=1.0.0

# 测试框架
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 开发工具
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# 文档生成
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
